requirements.txt
asyncio~=3.4.3
aiohttp~=3.8.4
asyncio-sse~=0.1
faster-whisper
grpcio~=1.54.2
grpcio-tools~=1.54.2
openai-whisper
pyaudio~=0.2.13
pyannote.audio~=2.1.1
setuptools==67.7.2
ssdpy~=0.4.1
torch==2.0.1
# whisperx @ git+https://github.com/m-bain/whisperx
numpy~=1.24.3
whisper~=1.1.10
